Create type for PostCreateInput, move database access to the service
