+++ src/modules/posts/types.ts
@@ -1,0 +1,3 @@
+import { Posts } from '@prisma/client';
+
+export type PostCreateInput = Pick<Posts, 'title' | 'content'> & { authorEmail: string };
