+++ src/modules/users/index.ts
@@ -1,4 +1,4 @@
-import usersRoutes from './routes-users';
+import usersRoutes from './users.routes';
 
 export default {
   usersRoutes
