+++ src/modules/users/users.routes.ts
@@ -1,8 +1,9 @@
-import core from 'express';
+import core from "express";
 
-import usersController from './users-controller';
+import usersController from "./users.controller";
 
 export default function (app: core.Express) {
   app.get("/users", usersController.index);
   app.post("/users", usersController.create);
-};
+}
+
